requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
opencv-python-headless>=4.9.0
numpy>=1.26.0
pyserial>=3.5
ultralytics>=8.2.0
